scripts/requirements.txt
nibabel
numpy
Pillow
requests
tqdm
